Unit tests UrlReaderUtil
