+++ src/main/java/com/polifono/util/UrlReaderUtil.java
@@ -1,0 +1,19 @@
+package com.polifono.util;
+
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.URL;
+
+public class UrlReaderUtil {
+
+    public static String readURL(URL url) throws IOException {
+        ByteArrayOutputStream baos = new ByteArrayOutputStream();
+        InputStream is = url.openStream();
+        int r;
+        while ((r = is.read()) != -1) {
+            baos.write(r);
+        }
+        return new String(baos.toByteArray());
+    }
+}
